# Viewer snapshot rendering mode