--- a/packages/public/@babylonjs/viewer-alpha/test/apps/web/published.html
+++ b/packages/public/@babylonjs/viewer-alpha/test/apps/web/published.html
@@ -1,0 +1,37 @@
+<!doctype html>
+<html>
+    <head>
+        <title>Babylon Viewer Demo</title>
+        <meta charset="UTF-8" />
+        <script type="module" src="https://unpkg.com/@babylonjs/viewer@preview/dist/babylon-viewer.esm.min.js"></script>
+        <style>
+            html,
+            body {
+                width: 100%;
+                height: 100%;
+                padding: 0;
+                margin: 0;
+                overflow: hidden;
+            }
+        </style>
+    </head>
+    <body ondragover="event.preventDefault()" ondrop="onDrop(event)">
+        <babylon-viewer engine="WebGPU" source="https://playground.babylonjs.com/scenes/BoomBox.glb"> </babylon-viewer>
+        <script>
+            const viewerElement = document.querySelector("babylon-viewer");
+            async function onDrop(event) {
+                const file = event.dataTransfer.files[0];
+                if (file) {
+                    event.preventDefault();
+                    await customElements.whenDefined("babylon-viewer");
+                    const reader = new FileReader();
+                    reader.onload = function (e) {
+                        const dataUri = e.target.result;
+                        viewerElement.source = dataUri;
+                    };
+                    reader.readAsDataURL(file);
+                }
+            }
+        </script>
+    </body>
+</html>