--- a/packages/tools/viewer-alpha/test/apps/web/index.html
+++ b/packages/tools/viewer-alpha/test/apps/web/index.html
@@ -65,7 +65,7 @@
 
     <body ondragover="event.preventDefault()" ondrop="onDrop(event)">
         <babylon-viewer
-            engine="WebGL"
+            engine="WebGPU"
             source="https://raw.githubusercontent.com/BabylonJS/Assets/master/meshes/ufo.glb"
             environment="../../../../../public/@babylonjs/viewer-alpha/assets/photoStudio.env"
             animation-speed="1.5"
@@ -92,7 +92,7 @@
             viewerElement.addEventListener("viewerready", (event) => {
                 viewerDetails = viewerElement.viewerDetails;
 
-                startHotSpotLinesRenderLoop();
+                //startHotSpotLinesRenderLoop();
                 console.log(`Viewer ready.`, viewerDetails);
             });
             viewerElement.addEventListener("modelchange", (event) => {