This change mostly is about using the new `SnapshotRenderingHelper` to enable WebGPUEngine snapshot rendering mode. Unfortunately there are some mysteries with the perf impact, most notably, when I was testing all this on my MacBook two weeks ago, I was seeing the frame rate on my test model (20x20x20 cubes) go from ~60fps to ~95fps (on my 120hz display). Now, I am seeing ~58fps -> 66fps, even if I checkout the same branch from two weeks ago (not just the version that is using `SnapshotRenderingHelper`). I have no explanation for this. The only thing I can think of is that I installed a MacOS update since then. Behavior is the same in Chrome and Edge. With an even heavier model (25x25x25 cubes), I am seeing these results:

### MacBook
- WebGL: 45fps
- WebGPU: 25fps
- WebGPU w/snapshot: 27fps

### Windows
- WebGL: 48fps
- WebGPU: 40fps
- WebGPU w/snapshot: 55fps

So snapshot rendering as is still improves perf for the viewer, but not as dramatically as when previously tested, and it's now worse than `Engine` (WebGL) on my MacBook, so it would be good to investigate more separately.

I included a couple other small changes in this PR as well:
- When ViewerElement parses the hotspots string, it splits on any amount of white space, not just a single white space (@CedricGuillemet).
- I changed the default engine for the Viewer test app to be WebGPU for now.
- Disabled the test hotspot in the Viewer test app by default since it kind of gets in the way when testing other models.
- Added a published.html test file to make it easy to test the published viewer package.