@bjsplat 2024-10-10
Snapshot stored with reference name:
refs/pull/15684/merge

Test environment: 
https://snapshots-cvgtc2eugrd3cgfd.z01.azurefd.net/refs/pull/15684/merge/index.html

To test a playground add it to the URL, for example:

https://snapshots-cvgtc2eugrd3cgfd.z01.azurefd.net/refs/pull/15684/merge/index.html#WGZLGJ#4600

Links to test babylon tools with this snapshot:

https://playground.babylonjs.com/?snapshot=refs/pull/15684/merge
https://sandbox.babylonjs.com/?snapshot=refs/pull/15684/merge
https://gui.babylonjs.com/?snapshot=refs/pull/15684/merge
https://nme.babylonjs.com/?snapshot=refs/pull/15684/merge

To test the snapshot in the playground with a playground ID add it after the snapshot query string:

https://playground.babylonjs.com/?snapshot=refs/pull/15684/merge#BCU1XR#0