@bjsplat 2024-10-10
Visualization tests for WebGPU (Experimental)

https://snapshots-cvgtc2eugrd3cgfd.z01.azurefd.net/refs/pull/15684/merge/testResults/webgpuplaywright/index.html