@bjsplat 2024-10-10
WebGL2 visualization test reporter:

https://snapshots-cvgtc2eugrd3cgfd.z01.azurefd.net/refs/pull/15684/merge/testResults/webgl2playwright/index.html